+++ R/expand.R
@@ -28,7 +28,7 @@ function(x, y, filename="", overwrite=FALSE, ...) {
 		if (is.vector(y)) {
 			if (length(y) <= 2) {
 				y <- round(y)
-				stopifnot(all(y > 0))
+				stopifnot(all(y >= 0))
 				adj <- rev(y) * res(x)
 				y <- as.vector(ext(x))
 				y[1] <- y[1] - adj[1]