+++ R/match.R
@@ -1,9 +1,9 @@
 # Author: Robert J. Hijmans
 # License GPL v3
 
-getFactTable <- function(x, table, sender="%in%") {
+getCatIDs <- function(x, table, sender="%in%") {
 	if (!is.factor(x)) {
-		error(sender, "Can only match character values of x is categorical")
+		error(sender, "Can only match character values if x is categorical")
 	}
 	if (nlyr(x) != 1) {
 		error(sender, "matching with character values is only supported for single layer SpatRaster")
@@ -11,17 +11,18 @@ getFactTable <- function(x, table, sender="%in%") {
 	d <- cats(x)[[1]]
 	levs <- levels(x)[[1]]
 	m <- na.omit(match(table, levs))
-	if (length(m) == 0) {
-		return(as.logical(x*0))
-	}
 	d[m,1]
 }
+
 
 setMethod("match", signature(x="SpatRaster"),
 	function(x, table, nomatch=NA, incomparables=NULL) {
 		table <- unique(table)
 		if (is.character(table)) {
-			table <- getFactTable(x, table, sender="match")
+			table <- getCatIDs(x, table, sender="match")
+			if (length(table) == 0) {
+				return(as.logical(x*0))
+			}
 		}
 		app(x, function(i) match(i, table, nomatch, incomparables))
 	}
@@ -33,7 +34,10 @@ setMethod("%in%", signature(x="SpatRaster"),
 
 		table <- unique(table)
 		if (is.character(table)) {
-			table <- getFactTable(x, table)
+			table <- getCatIDs(x, table)
+			if (length(table) == 0) {
+				return(as.logical(x*0))
+			}
 		}	
 		opt <- spatOptions("", FALSE, list())
 		x@ptr <- x@ptr$is_in(table, opt)